pyproject.toml
[project]
name = "strands-agentcore-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.1.1",
    "bedrock-agentcore-starter-toolkit>=0.2.5",
    "boto3>=1.42.14",
    "strands-agents-tools>=0.2.18",
    "strands-agents[sdk]>=1.20.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
]

[tool.setuptools]
py-modules = ["my_agent", "invoke_agent"]
